Split fallback media helpers
